# Conflicts:
#	pom.xml
#	src/com/lilithsthrone/controller/MainController.java
#	src/com/lilithsthrone/controller/eventListeners/tooltips/TooltipInformationEventListener.java
#	src/com/lilithsthrone/game/Game.java
#	src/com/lilithsthrone/game/character/GameCharacter.java
#	src/com/lilithsthrone/game/character/body/types/EyeType.java
#	src/com/lilithsthrone/game/character/body/valueEnums/CumProduction.java
#	src/com/lilithsthrone/game/character/npc/NPC.java
#	src/com/lilithsthrone/game/character/npc/dominion/Vanessa.java
#	src/com/lilithsthrone/game/character/npc/fields/Ziva.java
#	src/com/lilithsthrone/game/combat/moves/CMFetishAttack.java
#	src/com/lilithsthrone/game/combat/spells/Spell.java
#	src/com/lilithsthrone/game/dialogue/npcDialogue/unique/LumiDialogue.java
#	src/com/lilithsthrone/game/dialogue/story/CharacterCreation.java
#	src/com/lilithsthrone/game/dialogue/utils/EnchantmentDialogue.java
#	src/com/lilithsthrone/game/dialogue/utils/InventoryDialogue.java
#	src/com/lilithsthrone/game/dialogue/utils/OptionsDialogue.java
#	src/com/lilithsthrone/game/dialogue/utils/UtilText.java
#	src/com/lilithsthrone/game/inventory/CharacterInventory.java
#	src/com/lilithsthrone/game/inventory/enchanting/TFPotency.java
#	src/com/lilithsthrone/game/inventory/item/AbstractItem.java
#	src/com/lilithsthrone/game/inventory/item/AbstractItemType.java
#	src/com/lilithsthrone/game/sex/Sex.java
#	src/com/lilithsthrone/game/sex/positions/SexPosition.java
#	src/com/lilithsthrone/game/sex/positions/slots/SexSlotLyingDown.java
#	src/com/lilithsthrone/game/sex/sexActions/baseActionsMisc/GenericPositioning.java
#	src/com/lilithsthrone/game/sex/sexActions/baseActionsSelf/SelfFingerAnus.java
#	src/com/lilithsthrone/game/sex/sexActions/universal/DoggyStyleAndProneBone.java
#	src/com/lilithsthrone/main/Main.java
#	src/com/lilithsthrone/rendering/RenderingEngine.java
#	src/com/lilithsthrone/utils/Util.java

+++ src/com/lilithsthrone/controller/InventoryController.java
@@ -111,11 +111,11 @@ public class InventoryController {
 			((EventTarget) MainController.document.getElementById(id)).addEventListener("click", e->{
 				if (Main.game.getPlayer().getMoney()>0 && InventoryDialogue.getNPCInventoryInteraction() == InventoryInteraction.FULL_MANAGEMENT) {
 					if (InventoryDialogue.getInventoryNPC() == null) {
-						Main.game.getPlayerCell().getInventory().incrementMoney((int) Math.max(1, Main.game.getPlayer().getMoney()*0.01f));
+						Main.game.getPlayerCell().getInventory().incrementMoney((long) Math.max(1, Main.game.getPlayer().getMoney()*0.01f));
 					} else {
-						InventoryDialogue.getInventoryNPC().incrementMoney((int) Math.max(1, Main.game.getPlayer().getMoney()*0.01f));
+						InventoryDialogue.getInventoryNPC().incrementMoney((long) Math.max(1, Main.game.getPlayer().getMoney()*0.01f));
 					}
-					Main.game.getPlayer().incrementMoney((int) -Math.max(1, Main.game.getPlayer().getMoney()*0.01f));
+					Main.game.getPlayer().incrementMoney((long) -Math.max(1, Main.game.getPlayer().getMoney()*0.01f));
 					Main.game.setContent(new Response("", "", Main.game.getCurrentDialogueNode()));
 				}
 			}, false);
@@ -126,11 +126,11 @@ public class InventoryController {
 			((EventTarget) MainController.document.getElementById(id)).addEventListener("click", e->{
 				if (Main.game.getPlayer().getMoney()>0 && InventoryDialogue.getNPCInventoryInteraction() == InventoryInteraction.FULL_MANAGEMENT) {
 					if (InventoryDialogue.getInventoryNPC() == null) {
-						Main.game.getPlayerCell().getInventory().incrementMoney((int) Math.max(1, Main.game.getPlayer().getMoney()*0.1f));
+						Main.game.getPlayerCell().getInventory().incrementMoney((long) Math.max(1, Main.game.getPlayer().getMoney()*0.1f));
 					} else {
-						InventoryDialogue.getInventoryNPC().incrementMoney((int) Math.max(1, Main.game.getPlayer().getMoney()*0.1f));
+						InventoryDialogue.getInventoryNPC().incrementMoney((long) Math.max(1, Main.game.getPlayer().getMoney()*0.1f));
 					}
-					Main.game.getPlayer().incrementMoney((int) -Math.max(1, Main.game.getPlayer().getMoney()*0.1f));
+					Main.game.getPlayer().incrementMoney((long) -Math.max(1, Main.game.getPlayer().getMoney()*0.1f));
 					Main.game.setContent(new Response("", "", Main.game.getCurrentDialogueNode()));
 				}
 			}, false);
@@ -185,8 +185,8 @@ public class InventoryController {
 			if (MainController.document.getElementById(id) != null) {
 				((EventTarget) MainController.document.getElementById(id)).addEventListener("click", e->{
 					if (InventoryDialogue.getInventoryNPC().getMoney()>0 && InventoryDialogue.getNPCInventoryInteraction() == InventoryInteraction.FULL_MANAGEMENT) {
-						Main.game.getPlayer().incrementMoney((int) Math.max(1, InventoryDialogue.getInventoryNPC().getMoney()*0.01f));
-						InventoryDialogue.getInventoryNPC().incrementMoney((int) -Math.max(1, InventoryDialogue.getInventoryNPC().getMoney()*0.01f));
+						Main.game.getPlayer().incrementMoney((long) Math.max(1, InventoryDialogue.getInventoryNPC().getMoney()*0.01f));
+						InventoryDialogue.getInventoryNPC().incrementMoney((long) -Math.max(1, InventoryDialogue.getInventoryNPC().getMoney()*0.01f));
 						Main.game.setContent(new Response("", "", Main.game.getCurrentDialogueNode()));
 					}
 				}, false);
@@ -196,8 +196,8 @@ public class InventoryController {
 			if (MainController.document.getElementById(id) != null) {
 				((EventTarget) MainController.document.getElementById(id)).addEventListener("click", e->{
 					if (InventoryDialogue.getInventoryNPC().getMoney()>0 && InventoryDialogue.getNPCInventoryInteraction() == InventoryInteraction.FULL_MANAGEMENT) {
-						Main.game.getPlayer().incrementMoney((int) Math.max(1, InventoryDialogue.getInventoryNPC().getMoney()*0.1f));
-						InventoryDialogue.getInventoryNPC().incrementMoney((int) -Math.max(1, InventoryDialogue.getInventoryNPC().getMoney()*0.1f));
+						Main.game.getPlayer().incrementMoney((long) Math.max(1, InventoryDialogue.getInventoryNPC().getMoney()*0.1f));
+						InventoryDialogue.getInventoryNPC().incrementMoney((long) -Math.max(1, InventoryDialogue.getInventoryNPC().getMoney()*0.1f));
 						Main.game.setContent(new Response("", "", Main.game.getCurrentDialogueNode()));
 					}
 				}, false);
@@ -248,8 +248,8 @@ public class InventoryController {
 			if (MainController.document.getElementById(id) != null) {
 				((EventTarget) MainController.document.getElementById(id)).addEventListener("click", e->{
 					if (Main.game.getPlayerCell().getInventory().getMoney()>0 && InventoryDialogue.getNPCInventoryInteraction() == InventoryInteraction.FULL_MANAGEMENT) {
-						Main.game.getPlayer().incrementMoney((int) Math.max(1, Main.game.getPlayerCell().getInventory().getMoney()*0.01f));
-						Main.game.getPlayerCell().getInventory().incrementMoney((int) -Math.max(1, Main.game.getPlayerCell().getInventory().getMoney()*0.01f));
+						Main.game.getPlayer().incrementMoney((long) Math.max(1, Main.game.getPlayerCell().getInventory().getMoney()*0.01f));
+						Main.game.getPlayerCell().getInventory().incrementMoney((long) -Math.max(1, Main.game.getPlayerCell().getInventory().getMoney()*0.01f));
 						Main.game.setContent(new Response("", "", Main.game.getCurrentDialogueNode()));
 					}
 				}, false);
@@ -259,8 +259,8 @@ public class InventoryController {
 			if (MainController.document.getElementById(id) != null) {
 				((EventTarget) MainController.document.getElementById(id)).addEventListener("click", e->{
 					if (Main.game.getPlayerCell().getInventory().getMoney()>0 && InventoryDialogue.getNPCInventoryInteraction() == InventoryInteraction.FULL_MANAGEMENT) {
-						Main.game.getPlayer().incrementMoney((int) Math.max(1, Main.game.getPlayerCell().getInventory().getMoney()*0.1f));
-						Main.game.getPlayerCell().getInventory().incrementMoney((int) -Math.max(1, Main.game.getPlayerCell().getInventory().getMoney()*0.1f));
+						Main.game.getPlayer().incrementMoney((long) Math.max(1, Main.game.getPlayerCell().getInventory().getMoney()*0.1f));
+						Main.game.getPlayerCell().getInventory().incrementMoney((long) -Math.max(1, Main.game.getPlayerCell().getInventory().getMoney()*0.1f));
 						Main.game.setContent(new Response("", "", Main.game.getCurrentDialogueNode()));
 					}
 				}, false);
