+++ src/com/lilithsthrone/game/character/npc/fields/ElisAlleywayAttacker.java
@@ -160,7 +160,7 @@ public class ElisAlleywayAttacker extends NPC {
 
 	@Override
 	public void equipClothing(List<EquipClothingSetting> settings) {
-		this.incrementMoney((int) (this.getInventory().getNonEquippedValue() * 1f));
+		this.incrementMoney((long) (this.getInventory().getNonEquippedValue() * 1f));
 		this.clearNonEquippedInventory(false);
 		Main.game.getCharacterUtils().generateItemsInInventory(this, true, true, true);
