+++ src/com/lilithsthrone/game/character/npc/submission/RatGangMember.java
@@ -162,7 +162,7 @@ public class RatGangMember extends NPC {
 
 	@Override
 	public void equipClothing(List<EquipClothingSetting> settings) {
-		this.incrementMoney((int) (this.getInventory().getNonEquippedValue() * 0.5f));
+		this.incrementMoney((long) (this.getInventory().getNonEquippedValue() * 0.5f));
 		this.clearNonEquippedInventory(false);
 		
 		if(settings.contains(EquipClothingSetting.ADD_TATTOOS)) {
